requirements.txt
Cython==0.29.12
elasticsearch==7.7.0
elasticsearch-dsl==7.2.0
Flask==1.1.1
h5py==2.10.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
numpy==1.16.4
numpydoc==0.9.1
pandas==0.24.2
Pillow==6.1.0
pymilvus==0.2.11
scikit-image==0.15.0
scikit-learn==0.21.2
# scipy==1.3.0
tensorboard==2.2.1
tensorboard-plugin-wit==1.6.0.post3
# tensorflow==2.2.0
tensorflow-gpu==2.2.0
tensorflow-estimator==2.2.0
tornado==6.0.3
Werkzeug==0.15.4
ImageHash==4.1.0
